refactor(AnimePage): 💡 changes to use anilist on client
