+++ src/hooks/useNewestComments.ts
@@ -1,29 +1,49 @@
 import supabase from "@/lib/supabase";
+import { getMedia } from "@/services/anilist";
 import { Comment } from "@/types";
+import { MediaType } from "@/types/anilist";
 import { useSupabaseQuery } from "@/utils/supabase";
+import { useQuery } from "react-query";
 
 const mangaQuery = `
   *,
-  user:user_id(*),
-  manga:manga_id(id, title, vietnameseTitle)
+  user:user_id(*)
 `;
 
 const animeQuery = `
   *,
-  user:user_id(*),
-  anime:anime_id(id, title, vietnameseTitle)
+  user:user_id(*)
 `;
 
-const useNewestComments = <T extends "anime" | "manga">(type: T) => {
-  const isAnime = type === "anime";
+const useNewestComments = (type: MediaType) => {
+  const isAnime = type === MediaType.Anime;
 
-  return useSupabaseQuery(["newest-comments", type], () => {
-    return supabase
+  return useQuery(["newest-comments", type], async () => {
+    const { data, error } = await supabase
       .from<Comment>("comments")
       .select(isAnime ? animeQuery : mangaQuery)
       .not(isAnime ? "anime_id" : "manga_id", "is", null)
       .order("created_at", { ascending: false })
       .limit(15);
+
+    if (error) throw error;
+
+    const anilistMedia = await getMedia({
+      id_in: data.map((comment) =>
+        isAnime ? comment.anime_id : comment.manga_id
+      ),
+    });
+
+    return data.map((comment) => {
+      const media = anilistMedia.find(
+        (media) => media.id === (isAnime ? comment.anime_id : comment.manga_id)
+      );
+
+      return {
+        ...comment,
+        media,
+      };
+    });
   });
 };
