+++ src/utils/data.ts
@@ -1,6 +1,7 @@
 import enTranslations from "@/constants/en";
 import viTranslations from "@/constants/vi";
-import { Anime, Chapter, Episode, Manga, Media } from "@/types";
+import { Anime, Chapter, Episode, Manga } from "@/types";
+import { Media } from "@/types/anilist";
 import { parseNumbersFromString } from ".";
 
 type Translate = { readonly value: string; readonly label: string } & Record<
@@ -94,22 +95,18 @@ export const convert = (
   return constant[index].label;
 };
 
-export const getTitle = <T extends Media<Anime> | Media<Manga>>(
-  data: T,
-  locale?: string
-) => {
+export const getTitle = (data: Media, locale?: string) => {
   if (locale === "en") return data?.title.userPreferred;
 
   return data?.title.vietnamese || data?.title.userPreferred;
 };
 
-export const getDescription = <T extends Media<Anime> | Media<Manga>>(
-  data: T,
-  locale?: string
-) => {
-  if (locale === "en") return data?.description?.english;
+export const getDescription = (data: Media, locale?: string) => {
+  return data.description;
 
-  return data?.description?.vietnamese || data?.description?.english;
+  // if (locale === "en") return data?.description?.english;
+
+  // return data?.description?.vietnamese || data?.description?.english;
 };
 
 export const sortMediaUnit = <T extends Chapter | Episode>(data: T[]) => {
